+++ changelog.md
@@ -25,6 +25,7 @@
     - BREAKING: Renamed `createRipple()`'s `speed` prop values:
       - From `faster` to `fastest`
       - From `slower` to `slowest`
+    - BREAKING: `hasKey()` and `hasKeyFor()` now return a Promise, since they might need to load the given locale
     - `getCurrentMediaType()` now always returns `"video"` on YT instead of throwing an error
 
 </details> -->